update cart @Json verrsion 4

+++ src/main/java/com/poly/DATN_BookWorms/controller/CartController.java
@@ -58,7 +58,7 @@ public class CartController {
     }
 
     @RequestMapping("/delete/{cartid}")
-    public String deletecart(@PathVariable("cartid") Integer cartid) { 
+    public String deletecart(@PathVariable("cartid") Long cartid) { 
 //    	cartService.delete(cartid);
     	System.out.println("xmm");
     	cartService.delete(cartid);
