+++ src/main/java/com/poly/DATN_BookWorms/repo/BooksRepo.java
@@ -10,7 +10,7 @@ import com.poly.DATN_BookWorms.response.BookResponse;
 
 import java.util.List;
 
-public interface BooksRepo extends JpaRepository<Books, Integer> {
+public interface BooksRepo extends JpaRepository<Books, Long> {
 
     Books findFirstByOrderByQuantitysoldDesc();
 
@@ -35,7 +35,7 @@ public interface BooksRepo extends JpaRepository<Books, Integer> {
     Page<BookResponse> findCategoryBook(String category, Pageable pageable);
 
     @Override
-    List<Books> findAllById(Iterable<Integer> integers);
+    List<Books> findAllById(Iterable<Long> integers);
 
     @Query("SELECT b FROM Books b INNER JOIN b.listOfTypebooks tb WHERE tb.categories.categoryid = :categoryID")
     List<Books> findBooksByCategoryID(Integer categoryID);
